Remove the Leadership interface and just use a plain old Leadership class rather than a confusing LeadershipV2 class
